Big test cleanup.
reference_bare_repo fixture is now just that - it's a reference repo
to test git data retrieval. There is now a helper method to create a
test repository, which will just make a duplicate of the
temp_repo_template in the /tmp directory, all write tests are now done
using these temp repos instead. They have a small set of test objects
to work with also.