Ref packing support. Complete with a crazy bloated implementation using 2 mutexes. With an exception, I think I have a ref packing solution that will ensure existing loaded references, even if being used simultaneously to the packing process, synchronous or asynchronous, deleting/deleted or not, will still work.
The exception? I'm getting a race condition using one of the mutexes, 
I need to come up with a workaround, as I can't free the lock until
the object has been created, but I can't create the object until I'm
back on the main loop, which isn't feasible.

I'm worried I may have to overhaul the objectwrapping system a bit, so
I create the wrapping object anytime I want, and then wrap it in a V8
JS object when I hit the main loop. I actually just thought of that
while I was writing this git commit comment. The more I think about
it, that may be the solution, and it is a bit of an optimization too,
as I'll be able to load a git object, create a wrapping object for it,
load all the data, and then wrap it in a JS object when I come back to
main thread. That would eliminate the need to go into the EIO
threadpool a second time too, as well as meaning I can allocate the
C++ wrapping object in the EIO thread (also a minor improvement).

Will have to work on this tomorrow. Bedtime.